Seperated all functionality in controllers, services and routes.
