+++ routes/wishlist.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const router = express.Router();
+
+const wishlistController = require("../controllers/wishlistController");
+
+router.get("/getWishlist", wishlistController.getWishlist);
